Put WindowManager In Seperate File
